requirements/base.txt
Django
dj-database-url
django-configurations
django-extensions
whitenoise
